+++ src/index.ts
@@ -1,2 +1,1 @@
 export * from './zookeeperLock';
-export * from './exceptions';
